To Do List ver 1.1 without compile exe and jar
