+++ src/main/java/com/example/controller/constantsNotification/ErrorConstants.java
@@ -1,0 +1,14 @@
+package com.example.controller.constantsNotification;
+
+public class ErrorConstants {
+    public static final String ERROR_LOGIN_OR_PASSWORD_INCORRECT = "LOGIN OR PASSWORD INCORRECT";
+    public static final String ERROR_LOGIN_OR_SIGNUP_EMPTY_LOGIN_OR_PASSWORD = "YOU ARE TRYING TO ENTER AN EMPTY PASSWORD OR LOGIN";
+    public static final String ERROR_SAVE_EMPTY_TASK = "THE TASK CANNOT BE EMPTY";
+    public static final String ERROR_SIGN_UP_ACCOUNT_ALREADY_EXIST = "THIS ACCOUNT IS ALREADY EXIST";
+    public static final String ERROR_SIGN_UP_INVALID_CHARACTERS = "YOU ARE TRYING TO ENTER " +
+            "INVALID CHARACTERS. USE ONLY LATIN CHARACTERS FOR LOGIN AND PASSWORD";
+    public static final String ERROR_SIGN_UP_TO_LOONG_CHARACTERS = "YOU ARE TRYING TO ENTER " +
+            "A USERNAME OR PASSWORD THAT IS TOO LONG. " +
+            "THE LOGIN MUST NOT BE LONGER THAN 25 CHARACTERS, " +
+            "AND THE PASSWORD MUST NOT BE LONGER THAT 25 CHARACTERS ";
+}
