+++ src/main/java/com/example/controller/constantsNotification/SuccessfulConstants.java
@@ -1,0 +1,7 @@
+package com.example.controller.constantsNotification;
+
+public class SuccessfulConstants {
+    public static final String SUCCESSFUL_DELETE_ALL_USER = "ALL USERS HAVE BEEN DELETED AS WELL AS THEIR TASKS";
+    public static final String SUCCESSFUL_SIGN_UP = "ACCOUNT IS CREATED";
+
+}
