--- src/main/resources/DeniedLogin.fxml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="67.0" prefWidth="374.0" style="-fx-background-color: RED;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Text fill="WHITE" layoutY="45.0" strokeType="OUTSIDE" strokeWidth="0.0" text="LOGIN OR PASSWORD INCORRECT" textAlignment="CENTER" wrappingWidth="374.0">
-         <font>
-            <Font name="Agency FB Bold" size="30.0" />
-         </font>
-      </Text>
-   </children>
-</AnchorPane>
