--- src/main/resources/DeniedSignUpToLoongCharacters.fxml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="241.0" prefWidth="374.0" style="-fx-background-color: RED;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Text fill="WHITE" layoutY="40.0" strokeType="OUTSIDE" strokeWidth="0.0" text="YOU ARE TRYING TO ENTER A USERNAME OR PASSWORD THAT IS TOO LONG. THE LOGIN MUST NOT BE LONGER THAN 25 CHARACTERS, AND THE PASSWORD MUST NOT BE LONGER THAT 25 CHARACTERS " textAlignment="CENTER" wrappingWidth="374.0">
-         <font>
-            <Font name="Agency FB Bold" size="30.0" />
-         </font>
-      </Text>
-   </children>
-</AnchorPane>
