--- src/main/resources/SuccessfulDeleteAllUser.fxml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="85.0" prefWidth="374.0" style="-fx-background-color: GREEN;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Text fill="WHITE" layoutX="3.0" layoutY="36.0" strokeType="OUTSIDE" strokeWidth="0.0" text="ALL USERS HAVE BEEN DELETED AS WELL AS THEIR TASKS" textAlignment="CENTER" wrappingWidth="366.79998779296875">
-         <font>
-            <Font name="Agency FB Bold" size="31.0" />
-         </font>
-      </Text>
-   </children>
-</AnchorPane>
