+++ src/main/resources/SuccessfulNotification.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<AnchorPane fx:id="windowSuccessfulNotification" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="67.0" prefWidth="374.0" style="-fx-background-color: GREEN;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <Text fx:id="textSuccessfulNotification" fill="WHITE" layoutY="45.0" strokeType="OUTSIDE" strokeWidth="0.0" text="YOUR SUCCESSFUL" textAlignment="CENTER" wrappingWidth="374.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
+         <font>
+            <Font name="Agency FB Bold" size="31.0" />
+         </font>
+      </Text>
+   </children>
+</AnchorPane>
